qt: refactor GetDataThread, move to qt/util.py
